+++ R/hybrid.R
@@ -1,4 +1,4 @@
-#' @import parallel doParallel
+#' @import parallel doParallel snow
 NULL
 #> NULL
 
@@ -91,7 +91,50 @@ match_peaks <- function(aligned,
   return(pairing)
 }
 
-#' Add newly detected features to a known features table.
+
+#' A wrapper function to join knowledge from aligned features and known table.
+#' 
+#' @param features A list object with three tibble tables: metadata, intensity, and rt.
+#' @param known_table A table of known/previously detected peaks.
+#' @param match_tol_ppm The ppm tolerance to match identified features to known metabolites/features.
+#' @param mz_tol_relative The m/z tolerance level for peak alignment. The default is NA, which allows the program to search for the 
+#'  tolerance level based on the data. This value is expressed as the percentage of the m/z value. This value, multiplied by the m/z 
+#'  value, becomes the cutoff level.
+#' @param rt_tol_relative The retention time tolerance level for peak alignment. The default is NA, which allows the program to search for 
+#'  the tolerance level based on the data.
+#' @param from_features_to_known Determines direction of joining; if TRUE, aligned features are joined to known table, vice verse if it is FALSE.
+#' @param new_feature_min_count The number of profiles a new feature must be present for it to be added to the database.
+#' @return Enriched aligned table or known features.
+#' @export
+merge_features_and_known_table <- function(
+  features,
+  known_table,
+  match_tol_ppm,
+  mz_tol_relative,
+  rt_tol_relative,
+  from_features_to_known_table = TRUE,
+  new_feature_min_count = NA) {
+    if (from_features_to_known_table) {
+        return(augment_known_table(features,
+                                   known_table,
+                                   match_tol_ppm,
+                                   mz_tol_relative,
+                                   rt_tol_relative,
+                                   new_feature_min_count)
+               )
+    } else {
+        return(enrich_table_by_known_features(features,
+                                              known_table,
+                                              match_tol_ppm,
+                                              mz_tol_relative,
+                                              rt_tol_relative)
+               )
+    }
+}
+
+
+#' Add entries from the known features table to the aligned table.
+#' 
 #' @param aligned A list object with three tibble tables: metadata, intensity, and rt.
 #' @param known_table A table of known/previously detected peaks.
 #' @param match_tol_ppm The ppm tolerance to match identified features to known metabolites/features.
@@ -100,10 +143,9 @@ match_peaks <- function(aligned,
 #'  value, becomes the cutoff level.
 #' @param rt_tol_relative The retention time tolerance level for peak alignment. The default is NA, which allows the program to search for 
 #'  the tolerance level based on the data.
-#' @return Known table with novel features.
+#' @return Aligned table with known features.
 #' @import dplyr
-#' @export
-augment_with_known_features <- function(
+enrich_table_by_known_features <- function(
   aligned,
   known_table,
   match_tol_ppm,
@@ -135,6 +177,18 @@ augment_with_known_features <- function(
   return(aligned)
 }
 
+#' Add newly detected aligned features to a known features table.
+#' 
+#' @param aligned A list object with three tibble tables: metadata, intensity, and rt.
+#' @param known_table A table of known/previously detected peaks.
+#' @param match_tol_ppm The ppm tolerance to match identified features to known metabolites/features.
+#' @param mz_tol_relative The m/z tolerance level for peak alignment. The default is NA, which allows the program to search for the 
+#'  tolerance level based on the data. This value is expressed as the percentage of the m/z value. This value, multiplied by the m/z 
+#'  value, becomes the cutoff level.
+#' @param rt_tol_relative The retention time tolerance level for peak alignment. The default is NA, which allows the program to search for 
+#'  the tolerance level based on the data.
+#' @param new_feature_min_count The number of profiles a new feature must be present for it to be added to the database.
+#' @return Known table with novel features.
 augment_known_table <- function(
   aligned,
   known_table,
@@ -258,31 +312,42 @@ hybrid <- function(
 
   # NOTE: side effect (doParallel has no functionality to clean up)
   doParallel::registerDoParallel(cluster)
+  register_functions_to_cluster(cluster)
 
   check_files(filenames)
   sample_names <- get_sample_name(filenames)
   number_of_samples <- length(sample_names)
-
+  
   message("**** feature extraction ****")
-  extracted <- extract_features(
-    cluster = cluster,
-    filenames = filenames,
-    min_presence = min_pres,
-    min_elution_length = min_run,
-    mz_tol = mz_tol,
-    baseline_correct = baseline_correct,
-    baseline_correct_noise_percentile = baseline_correct_noise_percentile,
-    intensity_weighted = intensity_weighted,
-    min_bandwidth = min_bandwidth,
-    max_bandwidth = max_bandwidth,
-    sd_cut = sd_cut,
-    sigma_ratio_lim = sigma_ratio_lim,
-    shape_model = shape_model,
-    peak_estim_method = peak_estim_method,
-    component_eliminate = component_eliminate,
-    moment_power = moment_power,
-    BIC_factor = BIC_factor
-  )
+  profiles <- snow::parLapply(cluster, filenames, function(filename) {
+      proc.cdf(
+          filename = filename,
+          min_presence = min_pres,
+          min_elution_length = min_run,
+          mz_tol = mz_tol,
+          baseline_correct = baseline_correct,
+          baseline_correct_noise_percentile = baseline_correct_noise_percentile,
+          intensity_weighted = intensity_weighted,
+          do.plot = FALSE,
+          cache = FALSE
+      )
+  })
+  
+  extracted <- snow::parLapply(cluster, profiles, function(profile) {
+      prof.to.features(
+          profile = profile,
+          min.bw = min_bandwidth,
+          max.bw = max_bandwidth,
+          sd.cut = sd_cut,
+          sigma.ratio.lim = sigma_ratio_lim,
+          shape.model = shape_model,
+          estim.method = peak_estim_method,
+          component.eliminate = component_eliminate,
+          power = moment_power,
+          BIC.factor = BIC_factor,
+          do.plot = FALSE
+      )
+  })
 
  message("**** computing clusters ****")
   extracted_clusters <- compute_clusters(
@@ -325,12 +390,13 @@ hybrid <- function(
   )
 
   message("**** augmenting with known peaks ****")
-  merged <- augment_with_known_features(
-    aligned = aligned,
-    known_table = known_table,
-    match_tol_ppm = match_tol_ppm,
-    mz_tol_relative = adjusted_clusters$mz_tol_relative,
-    rt_tol_relative = adjusted_clusters$rt_tol_relative
+  merged <- merge_features_and_known_table(
+      features = aligned,
+      known_table = known_table,
+      match_tol_ppm = match_tol_ppm,
+      mz_tol_relative = adjusted_clusters$mz_tol_relative,
+      rt_tol_relative = adjusted_clusters$rt_tol_relative,
+      from_features_to_known_table = FALSE
   )
 
   message("**** weaker signal recovery ****")
@@ -399,12 +465,13 @@ hybrid <- function(
   )
 
   message("**** augmenting known table ****")
-  augmented <- augment_known_table(
-    aligned = recovered_aligned,
+  augmented <- merge_features_and_known_table(
+    features = recovered_aligned,
     known_table = known_table,
     match_tol_ppm = match_tol_ppm,
     mz_tol_relative = adjusted_clusters$mz_tol_relative,
     rt_tol_relative = adjusted_clusters$rt_tol_relative,
+    from_features_to_known_table = TRUE,
     new_feature_min_count = new_feature_min_count
   )